refactor(connectors): move google connector package to top-level workspace
