+++ connectors/google/src/docs/tools.ts
