+++ connectors/google/src/scopes.ts
