+++ packages/server/package.json
@@ -21,7 +21,7 @@
     "@hono/node-server": "^1.19.11",
     "@hono/zod-validator": "^0.7.6",
     "@openrouter/ai-sdk-provider": "catalog:ai",
-    "@stitch/google": "workspace:*",
+    "@stitch-connectors/google": "workspace:*",
     "@stitch/recordings": "workspace:*",
     "@stitch/shared": "workspace:*",
     "ai": "catalog:ai",
